requirements.txt
discord.py
mysql-connector-python
aiohttp
python-dotenv
Flask
Flask-SocketIO
waitress
google-generativeai
openai